Don't truncate floating point numbers

+++ src/main/java/com/easternedgerobotics/rov/math/MovingAverage.java
@@ -6,7 +6,7 @@ public final class MovingAverage {
     public static Observable<Float> from(final Observable<Float> source, final int n) {
         final RingBuffer buffer = new RingBuffer(n);
         return source.map(value -> {
-            buffer.add(value.longValue());
+            buffer.add(value);
             return buffer.apply(s -> (float) s.summaryStatistics().getAverage());
         });
     }
